Data containing conditional projections for the logarithm of GDP (gdp) for 189 United Nations countries from 2024 to 2029

For each of the countries a time series of 6 observations on GDP growth rates (gdp) formatted so they is provided to generate conditional forecasts of labour market outcomes given the provided projected paths of output. Last data update was implemented on 2024-09-21.

Usage

data(ilo_conditional_forecasts)

Format

A list of 189 ts objects with time series of 6 observations on 4 variables:

gdp

logarithm of gross domestic product - contains projected values

UR

unemployment rate - contains missing values

EPR

annual employment rate - contains missing values

LFPR

annual labour force participation rate - contains missing values

Source

International Labour Organization. (2020). ILO modelled estimates database, ILOSTAT [database]. Available from https://ilostat.ilo.org/data/.

Examples

data(ilo_conditional_forecasts)   # upload the data